realize the table of arguments
